Add a validation for a user entering their first and last name.
Feedback from the workshop: "Name and Last Names need to be mandatory:
We noticed that during our workshop some of the students who made a
request, some of them, only used their names and they didnt write their
last names, which is mandatory according to our Law when you make a
request. Is it possible to work and another thechnical reform or adjust
in the Register Part and to separate in two different boxs NAME and LAST
NAME?"

Using separate boxes for the first and last names was considered as an
alternative approach, but gets more complicated because we want
validation messages for the first and last name fields separately and
want to remove the validation message from the combined name attribute
that is coming from the core code. This is possible e.g using code as
suggested in
http://stackoverflow.com/questions/7545938/how-to-remove-validation-using-instance-eval-clause-in-rails
but doesn't seem as clean or maintainable. If the new validation is
sufficient, it seems preferable to a lot of complexity.